+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/CheckRequirementsAction.java
@@ -1,0 +1,48 @@
+package com.elmakers.mine.bukkit.action.builtin;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+import org.bukkit.configuration.ConfigurationSection;
+
+import com.elmakers.mine.bukkit.action.CompoundAction;
+import com.elmakers.mine.bukkit.api.action.ActionHandler;
+import com.elmakers.mine.bukkit.api.action.CastContext;
+import com.elmakers.mine.bukkit.api.requirements.Requirement;
+import com.elmakers.mine.bukkit.api.spell.SpellResult;
+import com.elmakers.mine.bukkit.utility.ConfigurationUtils;
+
+public class CheckRequirementsAction extends CompoundAction {
+    private Collection<Requirement> requirements;
+
+    @Override
+    public void prepare(CastContext context, ConfigurationSection parameters)
+    {
+        super.prepare(context, parameters);
+        requirements = new ArrayList<>();
+        Collection<ConfigurationSection> requirementConfigurations = ConfigurationUtils.getNodeList(parameters, "requirements");
+        if (requirementConfigurations != null) {
+            for (ConfigurationSection requirementConfiguration : requirementConfigurations) {
+                requirements.add(new Requirement(requirementConfiguration));
+            }
+        }
+    }
+
+    protected boolean isAllowed(CastContext context) {
+        return context.getController().checkRequirements(context, requirements) == null;
+    }
+
+    @Override
+    public SpellResult step(CastContext context) {
+        boolean allowed = isAllowed(context);
+        ActionHandler actions = getHandler("actions");
+        if (actions == null || actions.size() == 0) {
+            return allowed ? SpellResult.CAST : SpellResult.STOP;
+        }
+
+        if (!allowed) {
+            return SpellResult.NO_TARGET;
+        }
+        return startActions();
+    }
+}