+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add spawn tag API, to support MagicWorlds 1.6
  - Alter now works on flower pots (1.13 only)
  - Add some safety checks for class entity attributes getting stuck on server crash
+ - Add jump_cast option for brooms and vehicles, to cast a spell with spacebar while riding
 
 # 7.6.5